Split the tables components to individual files
